+++ packages/core/src/selectors/continuousListening.ts
@@ -1,0 +1,3 @@
+import type { ReduxState } from '../types/internal/ReduxState';
+
+export default ({ continuousListening }: ReduxState): boolean => continuousListening;